sliding window function improvement and inference results

+++ HumanActivityRecognition/inference_on_ball_sampler_macro.py
@@ -1,51 +1,89 @@
 import os
 import pandas as pd
 import numpy as np
-from app_config import REPORTS_DIR, FIGURES_DIR
+from app_config import REPORTS_DIR, FIGURES_DIR, MODELS_DIR
 import sliding_window_on_data
 from torch.utils.data import DataLoader
-import torch
-import torch.nn as nn
-from models.DeepConvLSTM import DeepConvLSTM, HARDataset, collate_fn, create_weighted_sampler
+import glob
+
+from models.DeepConvLSTM import DeepConvLSTM, HARDataset 
 import optuna
 import optuna.visualization as vis
 import train
+import torch
 import train_with_cm
 import matplotlib.pyplot as plt
+#definisco il path da cui leggere i .csv
+
+from utils.log_config import logger
 
 #definisco il path da cui leggere i .csv
 
 path='C:\codes\HumanActivityRecognition\data\pdd_data'
 print(path)
 
-
-
-
-
-
-
 #Visto che l'informazione relativa all'id del bambino lo abbiamo, così come abbiamo anche l'informazione relativa
 #al giocattolo, vado a filtrare i .csv in modo tale da avere solo le righe che hanno l'attività non nulla e poi
 #appendo tutte le righe delle righe non nulle in un unico dataframe
 #per tutti i file che terminano in .csv nella cartella path
+# Definisco il percorso della cartella contenente i CSV
 
-df_list_ball = [] #lista vuota per appendere i dataframe con attività non nulla
-for file in os.listdir(path):
-    if file.endswith('.csv'):
-        #leggo solo i file che dopo l'undescore ha BA
-        if file.split('_')[1]=='BA.csv':
+# Nome del file CSV finale
+final_csv_path = os.path.join(path, 'df_BA_non_null.csv')
+
+# Controllo se il file esiste già
+if os.path.exists(final_csv_path):
+    print(f"Il file {final_csv_path} esiste già. Lo sto caricando...")
+    df_ball = pd.read_csv(final_csv_path)
+else:
+    #trovo tutti i file che corrispondono a "BA" nella cartella path e li stampo a schermo
+    files = glob.glob(os.path.join(path, "*_BA*.csv"))
+    print("Files:", files)
+    
+    kid_ball, kid_ball_no_null = [], [] # liste per salvare utenti prima e dopo il merge 
+    df_list_ball = [] # lista vuota per appendere i dataframe con attività non nulla
+
+    for file in files:
+        df = pd.read_csv(file)
+        
+        print("Original shape:", df.shape)
+        kid_ball.append(df['kid_id'].unique())
+        df = df[df['action_id'] != 0] #filtro le righe con action_id non nullo
+        print("Filtered shape:", df.shape)
+        kid_ball_no_null.append(df['kid_id'].unique())
+
+        print("Columns:", df.columns)
+        print("Action counts:\n", df['action'].value_counts())
+        print("Toy counts:\n", df['toy_id'].value_counts())
+        print("="*50)  # Separatore tra i file
+
+    # mi stampo gli utenti prima di fare il merge e dopo il merge
+    logger.info(f"Numero di utenti che hanno fatto almeno un azione: {len(kid_ball_no_null)/len(kid_ball)}")
+
+    '''
+    Visto che l'informazione relativa all'id del bambino lo abbiamo, così come abbiamo anche l'informazione relativa
+    al giocattolo, vado a filtrare i .csv in modo tale da avere solo le righe che hanno l'attività non nulla e poi
+    appendo tutte le righe non nulle in un unico dataframe per tutti i file che terminano in .csv nella balltella path
+    '''
+
+    for file in os.listdir(path):
+        if not file.endswith('.csv'):
+            continue
+    
+        #leggo solo i file che dopo l'undescore ha BE*.csv
+        if file.split('_')[-1].startswith('C') and file.endswith('.csv'): #controllo che il file termini con .csv
             df_temp=pd.read_csv(os.path.join(path,file))
             df_temp = df_temp[df_temp['action_id'] != 0]
             df_list_ball.append(df_temp)
 
-df_ball = pd.concat(df_list_ball)
-print("Dimensioni del df_ball con tutte le attività non nulle")
-print(df_ball.shape)
-print(df_ball.columns)
-print(df_ball['action'].value_counts())
+    df_ball = pd.concat(df_list_ball)
+    print("Dimensioni del df_ball con tutte le attività non nulle")
+    print(df_ball.shape)
+    print(df_ball.columns)
+    print(df_ball['action'].value_counts())
 
-#salvo il dataframe
-df_ball.to_csv(os.path.join(path,'df_BA_non_null.csv'),index=False)
+    # salvo il dataframe
+    df_ball.to_csv(os.path.join(path,'df_BA_non_null.csv'),index=False)
 
 
 #ora divido il dataframe in base all'attività (action_id) e salvo i dataframe in un file .csv
@@ -53,72 +91,46 @@ df_ball.to_csv(os.path.join(path,'df_BA_non_null.csv'),index=False)
 
 for action_id in df_ball['action_id'].unique():
     df_action = df_ball[df_ball["action_id"] == action_id] #filtro il dataframe in base all'attività
-    print(f"Dimensioni del dataframe df_ball_action_{action_id}")
-    print(df_action.shape) #stampo le dimensioni del dataframe
-    print(df_action.columns) #stampo le colonne del dataframe
-    print(df_action['action'].value_counts()) #stampo il conteggio delle attività
+    logger.debug(f"Dimensioni del dataframe df_ball_action_{action_id} - {df_action.shape}") #log delle dimensioni del dataframe
+    logger.info(f"Conteggio delle attività per df_ball_action_{action_id}") #log del conteggio delle attività
     #salvo il dataframe
     df_action.to_csv(os.path.join(path,f'df_ball_action_{action_id}.csv'),index=False) #index=False per non salvare l'indice
-    print(f"Salvato il dataframe df_ball_action_{action_id}.csv")
-
+    logger.debug(f"Salvato il dataframe df_ball_action_{action_id}.csv")
 
 #applico sliding window con la funzion process_csv
-#definisco i parametri
-nb_sensor_channels = 13
+nb_sensor_channels = 9
 sliding_window_length = 100
-sliding_window_step = 20
+sliding_window_step = 50
 
-#ora applico la funzione sliding window (che mi da come output x_window e y_window) a tutti i .csv relativi al toy palla
-#e poi concateno tutto in un unica x_train, y_train
+#ora applico la funzione sliding window (che mi da come output x_window e y_window) a tutti i .csv relativi al giocattolo ball
+#e poi concateno tutto in un unica x e y 
 
-X= []
-Y= []
+X, Y = [], []
+kid_action_counts = {}
 
-for file in os.listdir(path):
-    if file.endswith('.csv') and file.split('_')[1] == 'ball':
-        file_path = os.path.join(path, file)
-        print(file_path)
-        X_windows, Y_windows = sliding_window_on_data.process_csv(file_path, nb_sensor_channels, sliding_window_length, sliding_window_step)
-        X.append(X_windows)
-        Y.append(Y_windows)
+for action_file in [f for f in os.listdir(path) if f.endswith('.csv') and f.split('_')[1] == 'ball']:
+    file_path = os.path.join(path, action_file)
+    action = action_file.split('_')[-1].split('.')[0]
 
-# Concatenate all the windows into a single array
+    X_windows, Y_windows, kid_id_action_dict = sliding_window_on_data.process_csv(file_path, nb_sensor_channels, sliding_window_length, sliding_window_step)
+
+    X.append(X_windows)
+    Y.append(Y_windows)
+
+
+    logger.info(f"Numero  totale di finestre per l'azione {action}:{len(X_windows)}")
+    kid_action_counts[f"ball_action_{action}"] = kid_id_action_dict #aggiungo il dizionario al dizionario principale per tenere traccia del numero di finestre per ogni bambino per ogni azione, ogni ball_action è una chiave e il valore è un dizionario con il numero di finestre per ogni bambino
+    logger.info(f"Contenuto finale di kid_action_counts: {kid_action_counts}") #per veere quante finestre per ogni azione e per ogni bambino sono state elaborte 
+
+
+# Concateno tutti i dati in un unico array per X e Y
 X = np.concatenate(X, axis=0)
 Y = np.concatenate(Y, axis=0)
 
-#NUMERO TOTALE DI FINESTRE PER LA PALLA
-print("Numero totale di finestre per il giocattolo ball:")
-print(X.shape)
-print(Y.shape)
 
-#verifca
-print(X)
-print(Y)
-
-
-# estraggo gli id dei bambini per vedere quanti ne ho
-kid_ids = X[:, :, -2]  # Assuming kid_id is the third last column
-
-# Get unique kid_ids
-unique_kid_ids = np.unique(kid_ids)
-print("Unique kid_ids in X:")
-print(unique_kid_ids)
-
-
-
-#Splitto il dataset in base al numero di azioni eseguite per avere congruenza temporale tra train e test e per 
-#cercare di bilanciare le finestre in train e test
-
-#filtro per ogni
-#filtro per contare il numero di finestre per ogni azione
-unique_actions, counts = np.unique(Y, return_counts=True)
-print(unique_actions)
-action_counts = dict(zip(unique_actions, counts))
-
-print("Numero di finestre per ogni azione:")
-for action, count in action_counts.items():
-    print(f"Azione {action}: {count} finestre")
-
+# Stampo le dimensioni di X e Y
+logger.info(f"Dimensioni di X finale: {X.shape}")
+logger.info(f"Dimensioni di Y finale: {Y.shape}")
 
 #split ratio  (70% nel train e 30% nel test)
 split_ratio = 0.7
@@ -129,7 +141,7 @@ Y_train = []
 X_test = []
 Y_test = []
 
-for action in unique_actions:
+"""for action in unique_actions:
 
     #trovo gli indici delle finestre corrispondenti a ciascuna azione
     action_indices = np.where(Y == action)[0]
@@ -315,4 +327,4 @@ test_loader = DataLoader(test_dataset, batch_size=best_batch_size, shuffle=False
 
 
 best_f1_score = train_with_cm.train(model, train_loader, test_loader, epochs=100,batch_size=best_batch_size, lr=best_lr, figure_name="model_ball_inference_with_sampler_macro", patience=7, f1_average='macro')
-print(f"Best F1 score: {best_f1_score}")
+print(f"Best F1 score: {best_f1_score}")"""
